+++ ubelt/util_import.py
@@ -30,7 +30,7 @@ IS_PY_GE_308 = sys.version_info[0:2] >= (3, 8)  # type: bool
 IS_PY_LT_314 = sys.version_info[0:2] < (3, 14)  # type: bool
 
 
-class PythonPathContext(object):
+class PythonPathContext:
     """
     Context for temporarily adding a dir to the PYTHONPATH.
 
@@ -651,41 +651,44 @@ def _importlib_import_modpath(modpath):  # nocover
 
 
 def _importlib_modname_to_modpath(modname):  # nocover
-    import importlib.util
-    spec = importlib.util.find_spec(modname)
-    modpath = spec.origin.replace('.pyc', '.py')
-    return modpath
-
-
-def _pkgutil_modname_to_modpath(modname):  # nocover
     """
-    faster version of :func:`_syspath_modname_to_modpath` using builtin python
-    mechanisms, but unfortunately it doesn't play nice with pytest.
-
-    Note:
-        pkgutil.find_loader is deprecated in 3.12 and removed in 3.14
+    faster version of :func:`_syspath_modname_to_modpath` using builtin
+    python mechanisms, but unfortunately it doesn't play nice with pytest.
 
     Args:
         modname (str): the module name.
 
+    Note:
+        This has a side effect when you import subpckages, it forces the parent
+        package itself to import!
+
     Example:
         >>> # xdoctest: +SKIP
-        >>> from ubelt.util_import import _pkgutil_modname_to_modpath
+        >>> from ubelt.util_import import _importlib_modname_to_modpath
         >>> modname = 'xdoctest.static_analysis'
-        >>> _pkgutil_modname_to_modpath(modname)
+        >>> _importlib_modname_to_modpath(modname)
         ...static_analysis.py
         >>> # xdoctest: +REQUIRES(CPython)
-        >>> _pkgutil_modname_to_modpath('_ctypes')
+        >>> _importlib_modname_to_modpath('_ctypes')
         ..._ctypes...
 
     Ignore:
-        >>> _pkgutil_modname_to_modpath('cv2')
+        >>> _importlib_modname_to_modpath('cv2')
+        >>> from ubelt.util_import import _importlib_modname_to_modpath
+        >>> from ubelt.util_import import _syspath_modname_to_modpath
+        >>> import timerit
+        >>> for _ in timerit(label='ours'):
+        >>>     _syspath_modname_to_modpath('xdoctest.static_analysis')
+        >>> for _ in timerit(label='importlib'):
+        >>>     _importlib_modname_to_modpath('xdoctest.static_analysis')
+        Timed ours for: 10 loops, best of 5
+            time per loop: best=20.237 ms, mean=20.244 ± 0.0 ms
+        Timed stdlib for: 445407 loops, best of 5
+            time per loop: best=387.000 ns, mean=424.680 ± 19.7 ns
     """
-    import pkgutil
-    loader = pkgutil.find_loader(modname)
-    if loader is None:
-        raise Exception('No module named {} in the PYTHONPATH'.format(modname))
-    modpath = loader.get_filename().replace('.pyc', '.py')
+    import importlib.util
+    spec = importlib.util.find_spec(modname)
+    modpath = spec.origin.replace('.pyc', '.py')  # is pyc replace needed anymore?
     return modpath
 
 
@@ -740,7 +743,6 @@ def modname_to_modpath(modname, hide_init=True, hide_main=False, sys_path=None):
         #     modpath = _importlib_modname_to_modpath(modname)
         # except Exception:
         #     modpath = _syspath_modname_to_modpath(modname, sys_path)
-        # modpath = _pkgutil_modname_to_modpath(modname, sys_path)
         modpath = _syspath_modname_to_modpath(modname, sys_path)
 
     if modpath is None:
@@ -1064,7 +1066,7 @@ def _parse_static_node_value(node):
         # value = dict(zip(keys, values))
     elif IS_PY_LT_314 and isinstance(node, (ast.NameConstant)):  # nocover
         value = node.value
-    elif isinstance(node, ast.Constant):
+    elif isinstance(node, ast.Constant):  # nocover
         value = node.value
     else:
         raise TypeError('Cannot parse a static value from non-static node '