+++ ubelt/util_stream.py
@@ -184,7 +184,7 @@ class TeeStringIO(io.StringIO):
         return super().flush()
 
 
-class CaptureStream(object):
+class CaptureStream:
     """
     Generic class for capturing streaming output from stdout or stderr
     """